+++ 02-EmulDEBUG.sh
@@ -10,4 +10,4 @@ fi
 echo ">>> Installing: $(basename "$apk")"
 adb -s emulator-5554 install -r "$apk"
 
-#sleep 2
+sleep 2